backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas>=2.2.0
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
firebase-admin==6.2.0
aiofiles==23.2.1
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1 